+++ crates/relayer/relayer-components/src/transaction/impls/poll_tx_response.rs
@@ -10,7 +10,7 @@ use hermes_runtime_components::traits::runtime::HasRuntime;
 use hermes_runtime_components::traits::sleep::CanSleep;
 use hermes_runtime_components::traits::time::HasTime;
 
-use crate::error::traits::retry::HasRetryableError;
+use crate::error::traits::HasRetryableError;
 use crate::transaction::traits::poll_tx_response::{TxResponsePoller, TxResponsePollerComponent};
 use crate::transaction::traits::query_tx_response::CanQueryTxResponse;
 use crate::transaction::traits::types::tx_hash::HasTransactionHashType;
